refactor: move vite configs in **infra** to target-state locations
