+++ src/common/loadable/internal/vite.config.ts
@@ -1,6 +1,6 @@
 import { mergeConfig } from 'vite';
-import baseViteConfig from '../../../../infra/build/packages/vite.config.base';
-import { viteConfigForPackage } from '../../../../infra/build/packages/vite.config.package';
+import baseViteConfig from '../../../../infra/build/vite.config.base';
+import { viteConfigForPackage } from '../../../../infra/build/vite.config.package';
 import pkg from './package.json';
 import { PackageJson } from 'type-fest';
