+++ vite.config.ts
@@ -1,3 +1,3 @@
-import viteConfigBase from './infra/build/packages/vite.config.base';
+import viteConfigBase from './infra/build/vite.config.base';
 
 export default viteConfigBase
